+++ src/s2/workers/process/line/postprocessLine.js
@@ -78,7 +78,7 @@ export default function postprocessLine (mapID: string, source: string, tileID: 
   }
 
   // Upon building the batches, convert to buffers and ship.
-  const vertexBuffer = new Float32Array(vertices).buffer
+  const vertexBuffer = new Int16Array(vertices).buffer
   const featureGuideBuffer = new Float32Array(featureGuide).buffer
   // ship the vector data.
   postMessage({
